% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_coralseed.R
\name{animate_coralseed}
\alias{animate_coralseed}
\title{Animate Coral Seed Particle Tracks}
\usage{
animate_coralseed(
  input,
  filename,
  width = 1200,
  height = 600,
  delay = 10,
  loop = FALSE
)
}
\arguments{
\item{input}{An sf object containing particle tracks. Must include a 'time' and 'id' column.}

\item{filename}{Path to the output gif file.}

\item{width}{Width of the output gif in pixels. Default is 1200.}

\item{height}{Height of the output gif in pixels. Default is 600.}

\item{delay}{Delay between frames in the gif, in milliseconds. Default is 10.}

\item{loop}{Logical; should the gif loop? Default is FALSE.}
}
\value{
Saves a gif to the specified filename and returns the animation object.
}
\description{
This function animates particle tracks over time using tmap, creating a .gif file.
}
\details{
Packages required: \code{tmap}, \code{sf}, \code{dplyr}
}
\examples{
# library(sf)
# library(dplyr)
# library(tmap)
#
# sf::sf_use_s2(FALSE)
# lizard_particles_raw <- sf::st_read("/path/to/file.json", quiet=TRUE) |>
#   dplyr::filter(id < 100) |>
#   sf::st_zm(drop = TRUE, what = "ZM")
#
# animate_coralseed(
#   input = lizard_particles_raw,
#   filename = "/path/to/save/test.gif"
# )

}
