% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthic_map.R
\docType{data}
\name{benthic_map}
\alias{benthic_map}
\title{benthic_map}
\format{
An object of class \code{sf}.
}
\source{
See index.html
}
\usage{
benthic_map
}
\description{
example benthic_map from Lizard Island
}
\examples{
data(benthic_map)

}
\keyword{datasets}
