% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthic_probability.R
\docType{data}
\name{benthic_probability}
\alias{benthic_probability}
\title{benthic_probability}
\format{
An object of class \code{df}.
}
\source{
See index.html
}
\usage{
benthic_probability
}
\description{
example benthic_probability matrix of settlement types for habitat classes (mean and se)
}
\examples{
data(benthic_probability)

}
\keyword{datasets}
