% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dispersion.R
\name{calculate_dispersion}
\alias{calculate_dispersion}
\title{Calculate dispersion}
\usage{
calculate_dispersion(input)
}
\arguments{
\item{input}{sf POINTS file (defaults to NULL)}
}
\description{
Function to calculate dispersion of settled points
}
\details{
dispersion = standard deviation of distances from each point to centroid of all points
}
