% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralseed-package.R
\docType{package}
\name{coralseed-package}
\alias{coralseed}
\alias{coralseed-package}
\title{coralseed: Spatially Explicit Reef-Scale Model of Larval Reseeding}
\description{
A spatially explicit, individual-based probabilistic model for predicting coral settlement following larval dispersal on coral reefs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marine-ecologist.github.io/coralseed/}
  \item Report bugs at \url{https://github.com/marine-ecologist/coralseed/issues}
}

}
\author{
\strong{Maintainer}: George Roff \email{george.roff@csiro.au} (\href{https://orcid.org/0000-0002-2823-0575}{ORCID})

}
\keyword{internal}
