% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement_heatmap.R
\name{.settler_heatmap_legacy}
\alias{.settler_heatmap_legacy}
\title{Generate a Heatmap of Settler Particles}
\usage{
.settler_heatmap_legacy(
  input,
  xres = 20,
  yres = 20,
  buffer_factor = 0.1,
  n = 100,
  threshold = 10
)
}
\arguments{
\item{input}{An sf object containing point geometries representing settler particles.}

\item{xres}{Numeric. The resolution of the heatmap in the x direction meters. Default is 20.}

\item{yres}{Numeric. The resolution of the heatmap in the y direction meters. Default is 20.}

\item{buffer_factor}{Numeric. The proportion of the bounding box range to expand the heatmap extent. Default is 0.1 10\%.}

\item{n}{KDE grid resolution}

\item{threshold}{Numeric. Thresholding value < will remove low-density values}
}
\description{
This function computes a kernel density estimate (KDE) heatmap from spatial
points (settler particles) and returns a raster representation. The bounding
box is expanded by a specified factor to avoid edge effects.
}
\keyword{internal}
