% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parallel.R
\name{extract_parallel}
\alias{extract_parallel}
\title{Settlement statistics}
\usage{
extract_parallel(input, type, ...)
}
\arguments{
\item{input}{list of \%dopar\% outputs}

\item{...}{pass arguments}
}
\description{
Function to extract points and paths from multiple \%dopar\% outputs
}
