% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart_coralseed.R
\name{flowchart_coralseed}
\alias{flowchart_coralseed}
\title{Flowchart of Coral Larval Fates}
\usage{
flowchart_coralseed(
  seed_particles_input,
  settle_particles_input,
  save_output = NULL,
  multiplier = 1,
  postsettlement = 0.8,
  width = 10,
  height = 6
)
}
\arguments{
\item{seed_particles_input}{An \code{sf} object or tibble containing all released larval trajectories with at least \code{id}, \code{dispersaltime}, and \code{state}.}

\item{settle_particles_input}{A named list containing \verb{$points}, an \code{sf} object with settled larvae and \code{id}.}

\item{multiplier}{Numeric multiplier to scale particle counts (e.g., if each simulated particle represents multiple larvae).}

\item{postsettlement}{Postsettlement mortality in proportion (e.g. 80\% = 0.8)}
}
\value{
An interactive Sankey diagram (htmlwidget).
}
\description{
Generate an interactive Sankey diagram showing the fate of coral larvae
through dispersal, settlement, and mortality.
}
\examples{
\dontrun{
flowchart_coralseed(moore_particles, moore_settlers, multiplier = 1, postsettlement = 0.8)
}



}
