% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_silent.R
\name{ggplot_silent}
\alias{ggplot_silent}
\title{Silence ggplot warnings}
\usage{
ggplot_silent(...)
}
\arguments{
\item{...}{arguments are passed to \code{ggplot2:::print.ggplot()}}
}
\description{
replacement for ggplot() call that mutes warnings
}
