% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zarr.R
\name{import_zarr}
\alias{import_zarr}
\title{Seed particles}
\usage{
import_zarr(input, origin = "1970-01-01 15:00:00", crs = 4326)
}
\arguments{
\item{input}{input zarr folder}

\item{crs}{coordinate reference system (default = 4326)}
}
\description{
Function to convert xarrr arrays from oceanparticles to df format
}
\details{
zarr_file_path <- "/Users/rof011/oceanparcels/outputs/Elford_16073_Slope_71a.zarr/"
}
