% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infamis_tiles_exp.R
\docType{data}
\name{infamis_tiles_exp}
\alias{infamis_tiles_exp}
\title{A fitted brms model for tile-level settlement}
\format{
A \code{brmsfit} object with N chains, etc.
}
\source{
Simulated or observed coral settlement data.
}
\usage{
infamis_tiles_exp
}
\description{
A fitted brms model for tile-level settlement
}
\keyword{datasets}
