% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infamis_tiles_log}
\alias{infamis_tiles_log}
\title{Infamis tiles (lognormal model)}
\format{
A list or data frame of model parameters.
}
\source{
Fitted from infamis dispersal simulation outputs.
}
\usage{
infamis_tiles_log
}
\description{
Pre-computed infamis tile parameters fitted with a lognormal distribution.
}
\keyword{datasets}
