% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infamis_tiles_weibull}
\alias{infamis_tiles_weibull}
\title{Infamis tiles (Weibull model)}
\format{
A list or data frame of model parameters.
}
\source{
Fitted from infamis dispersal simulation outputs.
}
\usage{
infamis_tiles_weibull
}
\description{
Pre-computed infamis tile parameters fitted with a Weibull distribution.
}
\keyword{datasets}
