% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizard_osm_map}
\alias{lizard_osm_map}
\title{Lizard Island OSM basemap}
\format{
A basemap object compatible with tmap.
}
\source{
OpenStreetMap contributors.
}
\usage{
lizard_osm_map
}
\description{
OpenStreetMap basemap tiles for the Lizard Island region.
}
\keyword{datasets}
