% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coralseed.R
\name{map_coralseed}
\alias{map_coralseed}
\title{Map coralseed settlers}
\usage{
map_coralseed(
  seed_particles = particles,
  settle_particles = settlers,
  seascape_probability = seascape,
  restoration.plot = c(100, 100)
)
}
\arguments{
\item{seed_particles}{input from seed_particles}

\item{settle_particles}{input from settle_particles}

\item{seascape_probability}{input from seascape_probability}

\item{restoration.plot}{dimensions of the restoration plot in metres}
}
\description{
Function to quickly map settlers
}
