% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coralseed_future.R
\name{map_coralseed_future}
\alias{map_coralseed_future}
\title{Map coralseed settlers subfunction}
\usage{
map_coralseed_future(
  points = points,
  paths = paths,
  seascape_probability = seascape,
  restoration.plot = c(100, 100)
)
}
\arguments{
\item{seascape_probability}{input from seascape_probability}

\item{restoration.plot}{dimensions of the restoration plot in metres}

\item{seed_particles}{input from seed_particles}

\item{settle_particles}{input from settle_particles}
}
\description{
Function to quickly map settlers for multiple
}
