% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{
An object of class \code{sf}.
}
\source{
See index.html
}
\usage{
metadata
}
\description{
Metadata for the five example dispersal datasets datasets
}
\examples{
data(metadata)

}
\keyword{datasets}
