% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_draws_exp.R
\docType{data}
\name{parameter_draws_exp}
\alias{parameter_draws_exp}
\title{parameter_draws_exp}
\format{
An object of class \code{df}.
}
\source{
See index.html
}
\usage{
parameter_draws_exp
}
\description{
example posterior draws for exponential brms model for time-to-competency
}
\examples{
data(parameter_draws_exp)

}
\keyword{datasets}
