% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_draws_log.R
\docType{data}
\name{parameter_draws_log}
\alias{parameter_draws_log}
\title{parameter_draws_log}
\format{
An object of class \code{df}.
}
\source{
See index.html
}
\usage{
parameter_draws_log
}
\description{
example posterior draws for lognormal brms model for time-to-competency
}
\examples{
data(parameter_draws_log)

}
\keyword{datasets}
