% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_draws_weibull.R
\docType{data}
\name{parameter_draws_weibull}
\alias{parameter_draws_weibull}
\title{parameter_draws_weibull}
\format{
An object of class \code{df}.
}
\source{
See index.html
}
\usage{
parameter_draws_weibull
}
\description{
example posterior draws for Weibull brms model for time-to-competency
}
\examples{
data(parameter_draws_weibull)

}
\keyword{datasets}
