% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_distances.R
\name{particle_distances}
\alias{particle_distances}
\title{Particle distances}
\usage{
particle_distances(input = NULL, tmax = NULL, type = "df")
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{tmax}{limit particle times to less than tmax}

\item{type}{export format to "sf" or "df" (see above)}

\item{...}{pass arguments}
}
\description{
Function to calculate distances of particles at any timepoint
limit the particle distances to tmax (in minutes, e.g. 60 for 1hr)
returns either  "sf" ("MULTILINGSTRING" for each id) or "df" (data.frame with total length for each id) depending on type="df" or type="sf"
}
