% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_draws.R
\name{plot_draws}
\alias{plot_draws}
\title{plot draws}
\usage{
plot_draws(input, slice = 100)
}
\arguments{
\item{input}{output from return_draws()}

\item{slice}{number of simulations to draw for spaghetti plot (default 100)}
}
\description{
Function to plot spaghetti draws from time-to-competency brms model
}
