% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_draws.R
\name{plot_draws_median}
\alias{plot_draws_median}
\title{plot draws}
\usage{
plot_draws_median(input)
}
\arguments{
\item{n}{number of simulations}
}
\description{
Function to plot median and 90\% CI draws from time-to-competency brms model
}
