% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_particles.R
\name{plot_particles}
\alias{plot_particles}
\title{Plot output from settle_particles}
\usage{
plot_particles(
  input = NULL,
  seascape = NULL,
  parameter = "dispersaltime",
  pch = 16,
  size = 0.5,
  ...
)
}
\arguments{
\item{input}{input}

\item{seascape}{seascape habitat input}

\item{parameter}{parameter (column from coralseed)}

\item{pch}{pch}

\item{size}{size}

\item{...}{passes functions}
}
\description{
Function to plot output of seed_particles in base R for quick cheks
}
