% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recom.R
\name{plot_recom}
\alias{plot_recom}
\title{plot_recom: Visualize Ocean Currents and Wind from NetCDF Data}
\usage{
plot_recom(
  nc_file = "/Users/rof011/GBR_connectivity/Oceanparcels/Moore_uv_2015-12-05.nc",
  parcels_reefs = sf::st_read("/Users/rof011/GBR_connectivity/inputs/Moore_2D.gpkg",
    quiet = TRUE),
  parcels_polygon =
    dplyr::filter(sf::st_read("/Users/rof011/GBR_connectivity/inputs/Moore_2D.gpkg",
    quiet = TRUE), site_id == "Moore_16071_Slope_39a")
)
}
\arguments{
\item{nc_file}{Character. Path to the NetCDF file containing oceanographic data.}

\item{parcels_reefs}{sf object. Spatial dataset of reef locations.}

\item{parcels_polygon}{sf object. Spatial dataset for a specific study site polygon.}
}
\value{
A combined ggplot object displaying temporal trends of ocean currents and wind,
along with spatial maps of current speeds at the slack current time.
}
\description{
Reads a NetCDF file containing ocean current and wind data, extracts relevant
variables at a specified depth (-2.35m), and generates time-series and spatial plots.
Highlights slack current time and visualizes magnitude and direction of currents and wind.
}
