% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_settlers.R
\name{plot_settlers}
\alias{plot_settlers}
\title{Plot output from settle_particles}
\usage{
plot_settlers(input = NULL, pch = 16, size = 0.5, ...)
}
\arguments{
\item{input}{input}

\item{pch}{pch}

\item{size}{size}

\item{...}{passes functions}
}
\description{
Function to plot output of seed_particles in base R for quick cheks
}
