% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_competency.R
\name{predict_competency}
\alias{predict_competency}
\title{Predict competency}
\usage{
predict_competency(
  input,
  n_particles = 100,
  max.time = NULL,
  seed.value = NULL,
  return.plot = FALSE,
  ...
)
}
\arguments{
\item{input}{brmsfit object from a model using 'lognormal', 'exponential', or 'weibull' family}

\item{n_particles}{number of random simulations}

\item{max.time}{optional maximum dispersal time limit (default is max time in input$data)}

\item{seed.value}{random seed for reproducibility}

\item{return.plot}{if TRUE, returns ggplot of survival curves + point estimates}

\item{...}{additional arguments (not used)}
}
\description{
Function to generate time-to-settlement probability distributions using a \code{brms} model
}
\details{
Examples:
predict_competency(infamis_tiles_exp, n_particles = 100)
predict_competency(infamis_tiles_weibull, n_particles = 100)
predict_competency(infamis_tiles_log, n_particles = 100)
}
