% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reef_map.R
\docType{data}
\name{reef_map}
\alias{reef_map}
\title{reef_map}
\format{
An object of class \code{sf}.
}
\source{
See index.html
}
\usage{
reef_map
}
\description{
example reef_map from Lizard Island
}
\examples{
data(reef_map)

}
\keyword{datasets}
