% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_draws.R
\name{return_draws}
\alias{return_draws}
\title{return draws}
\usage{
return_draws(input, fit, tmax = 12, by = 1, n = 1000, ...)
}
\arguments{
\item{input}{brms model fit}

\item{tmax}{max time in hours}

\item{by}{seq by in minutes}

\item{n}{number of simulations}
}
\description{
Function to return draws from time-to-competency brms model
}
