% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2.R
\name{save2}
\alias{save2}
\title{Safely Save R Objects Without Overwriting the File}
\usage{
save2(..., file = stop("'file' must be specified"))
}
\arguments{
\item{...}{Data frame(s), other object(s), and further arguments,
passed on to \code{\link[base]{save}}. Required.}

\item{file}{Path/file name to for output. Required.}
}
\description{
Wraps \code{\link[base]{save}}. If the file already exists,
it will not be overwritten. A message will be printed indicating whether
the file was successfully written or if it already existed.
}
\details{
code from https://github.com/stopsack/khsmisc/blob/HEAD/R/datahandling.R
}
