% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape_color_pal.R
\name{seascape_color_pal}
\alias{seascape_color_pal}
\title{Seascape color palette}
\usage{
seascape_color_pal()
}
\value{
A named character vector of colors
}
\description{
Returns a named vector of colors for coral seascape habitat classes.
}
