% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasims_tiles_exp}
\alias{seasims_tiles_exp}
\title{Seasims tiles (exponential model)}
\format{
A list or data frame of model parameters.
}
\source{
Fitted from seasims dispersal simulation outputs.
}
\usage{
seasims_tiles_exp
}
\description{
Pre-computed seasims tile parameters fitted with an exponential distribution.
}
\keyword{datasets}
