% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasims_tiles_log}
\alias{seasims_tiles_log}
\title{Seasims tiles (lognormal model)}
\format{
A list or data frame of model parameters.
}
\source{
Fitted from seasims dispersal simulation outputs.
}
\usage{
seasims_tiles_log
}
\description{
Pre-computed seasims tile parameters fitted with a lognormal distribution.
}
\keyword{datasets}
