% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_futures2.R
\name{seed_futures2}
\alias{seed_futures2}
\title{Seed futures}
\usage{
seed_futures2(
  input = NULL,
  example = NULL,
  seascape = NULL,
  simulate.mortality = "none",
  simulate.mortality.n = 0.1,
  competency.function = "exp",
  limit.time = NA,
  probability = "additive",
  tracks = FALSE,
  set.seed = NULL,
  ...
)
}
\arguments{
\item{input}{input}

\item{example}{example input}

\item{seascape}{shp file inputs from seascape_probability()}

\item{simulate.mortality}{set mortality type via simulate_mortality() one of "typeI","typeII", "typeIII" (defaults to "none")}

\item{simulate.mortality.n}{set proportion of corals to kill over a 24hr period, where 0 is none, 1 is 100 (defaults to 0.1 or 10\%)}

\item{competency.function}{set distribution to define competency from time-to-settlement model, one of "weibull", "exp", "log" (defaults to "exp")}

\item{limit.time}{limit the time series, for example 720 will limit the settlement results between 0-12hrs (defaults to "NA")}

\item{probability}{one of "additive", "rapid", "lagged")}

\item{tracks}{generate particle track outputs (multilinestrings)}

\item{set.seed}{set seed for consistent results (defaults to NULL)}

\item{...}{passes functions}
}
\description{
Combined function to seed particles without returning extra data
}
\details{
seed_particles() uses predict_competency(), simulate_mortality()
}
