% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_multicore_temp.R
\name{seed_multicore_temp}
\alias{seed_multicore_temp}
\title{Seed futures}
\usage{
seed_multicore_temp(input = NULL, set.seed = NULL, limit.time = NA, ...)
}
\arguments{
\item{input}{input}

\item{set.seed}{set seed for consistent results (defaults to NULL)}

\item{limit.time}{limit the time series, for example 720 will limit the settlement results between 0-12hrs (defaults to "NA")}

\item{...}{passes functions}
}
\description{
Combined function to seed particles without returning extra data
}
\details{
seed_particles() uses predict_competency(), simulate_mortality()
}
