% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_buffer.R
\name{set_buffer}
\alias{set_buffer}
\title{Set buffer}
\usage{
set_buffer(input = NULL, buffer = 150)
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{buffer}{distance surrounding particle boundary (defaults to 250m)}
}
\description{
Function to generate an sf file for bounding box. Used only for tmap visualisations
}
