% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restoration_plot.R
\name{set_restoration_plot}
\alias{set_restoration_plot}
\title{Set restoration plot}
\usage{
set_restoration_plot(input = NULL, width = NULL, length = NULL, center = NULL)
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{width}{width in metres}

\item{length}{length in metres}

\item{center}{not yet defined, but later to change centre point}
}
\description{
Function to create a rectangular polygon around the release point (centroid or t0 of particle release) for plotting in tmap
}
