% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restoration_plot_futures.R
\name{set_restoration_plot_futures}
\alias{set_restoration_plot_futures}
\title{Set restoration plot for futures}
\usage{
set_restoration_plot_futures(
  input = NULL,
  width = NULL,
  length = NULL,
  center = NULL
)
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{width}{width in metres}

\item{length}{length in metres}

\item{center}{not yet defined, but later to change centre point}
}
\description{
Function to create a rectangular polygon around the release point (centroid or t0 of particle release) for plotting in tmap
Ignores dispersaltime from set_restoration_plot
}
