% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settle_particles.R
\name{settle_particles}
\alias{settle_particles}
\title{Settle Particles}
\usage{
settle_particles(
  input,
  seascape = seascape,
  probability = "additive",
  silent = TRUE,
  return.plot = FALSE,
  subsample = NULL,
  seed.value = NULL,
  ...
)
}
\arguments{
\item{input}{output from \code{seed_particles()} or directly a particle \code{sf} object}

\item{seascape}{sf object from \code{seascape_probability()}, for plotting base}

\item{probability}{settlement rule, one of "additive", "rapid", or "lagged"}

\item{silent}{suppress printed output (default TRUE)}

\item{return.plot}{whether to return tmap plot output}

\item{subsample}{optionally sample N particles for faster plotting}

\item{seed.value}{set seed for reproducibility}

\item{...}{additional args}
}
\description{
Function to determine probability of settlement if particles pass suitable substrate.
Options:
\itemize{
\item "additive": settle somewhere in habitat intersect regardless of time
\item "lagged": sample first 10 minutes of habitat entry
\item "rapid": settle at first habitat intersect
}
}
