% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement_density.R
\name{settlement_density}
\alias{settlement_density}
\title{Settlement density}
\usage{
settlement_density(
  input = NULL,
  combined = FALSE,
  cellsize = 20,
  concavity = 1.2,
  length_threshold = 10,
  ...
)
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{combined}{test to allow multiple coralseed file inputs in list (defaults to FALSE)}

\item{cellsize}{dimensions of grid to count settlers (defaults to 20m)}

\item{concavity}{concavity value passed to concaveman (defaults to 2}

\item{length_threshold}{length threshold passed to concaveman (defaults to 0)}

\item{...}{pass arguments}
}
\description{
Function to calculate settlement density across a grid
Uses sp::over instead of sf for speed
}
