% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement_statistics.R
\name{settlement_statistics}
\alias{settlement_statistics}
\title{Settlement statistics}
\usage{
settlement_statistics(
  input = NULL,
  combined = FALSE,
  cellsize = 20,
  concavehull = TRUE,
  concavity = 2,
  length_threshold = 0,
  ...
)
}
\arguments{
\item{input}{input (defaults to NULL)}

\item{combined}{test to allow multiple coralseed file inputs in list (defaults to FALSE)}

\item{cellsize}{dimensions of grid to count settlers (defaults to 20m)}

\item{concavehull}{return concave hull around settled particles (via concaveman)}

\item{concavity}{concavity value passed to concaveman (defaults to 2}

\item{length_threshold}{length threshold passed to concaveman (defaults to 0)}

\item{...}{pass arguments}
}
\description{
Function to generate settlement statistics from coralseed output
}
