% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{Simulate Mortality}
\usage{
simulate_mortality(
  input = NULL,
  simulate.mortality = NULL,
  simulate.mortality.n = 0.1,
  return.plot = TRUE,
  silent = FALSE,
  set.seed = NULL,
  ...
)
}
\arguments{
\item{input}{input (defaults to particle_points_expanded from seed_particles() function)}

\item{simulate.mortality}{select which type of mortality (one of "typeI", "typeII", "typeIII")}

\item{simulate.mortality.n}{proportion of the population to kill (0 = none, 1 = 100\% mortality)}

\item{return.plot}{show plot of the results (defaults to FALSE)}

\item{silent}{silence printing results while running (defaults to FALSE)}

\item{set.seed}{set seed for consistent results (defaults to NULL)}

\item{...}{pass arguments}
}
\description{
Function to simulate mortality of a population. simulate.mortality.n is the proportion of the population to simulate mortality over a 24hr period.
}
\details{
Currently set to mortality types I,II,III parameterised from the weibull distribution (return.plot=TRUE to visualise results against 1000 simulations)

tmp <- simulate_mortality(particle_points_expanded, simulate.mortality="typeIII",  simulate.mortality.n = 0.2, return.plot=TRUE)
}
