% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_set_point.R
\name{st_set_point}
\alias{st_set_point}
\title{st_set_point}
\usage{
st_set_point(lon, lat, crs = 20353)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{crs}{coordinate reference system}
}
\description{
Function to generate sf point from a lon, lat, and a crs
}
