% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_particles.R
\name{stationary_particles}
\alias{stationary_particles}
\title{Stationary particles}
\usage{
stationary_particles(input)
}
\arguments{
\item{input}{input particle tracks (e.g. coralseed:::Mermaid_PointSource_Bay_01)}

\item{subset}{n particles to subset

stat_particles_liz22_del_14_33 <- stationary_particles(liz22_del_14_33)

stat_liz22_del_14_33_map <- stationary_particles_map(liz22_del_14_33 |> filter(id \%in\% seq(0:1000)))
stat_liz22_del_14_33_map <- stationary_particles_map(coralseed:::Mermaid_PointSource_Bay_01 |> filter(id \%in\% seq(0:1000)))

stat_liz22_del_14_33_map <- stationary_particles_map(run_day_12039_liz22_del_11_35|> filter(id \%in\% seq(0:100)))
stat_liz22_del_14_33_stationary <- stationary_particles(run_day_12039_liz22_del_11_35)}
}
\description{
Function to identify stationary particles in particle tracks
}
