% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_particles_map.R
\name{stationary_particles_map}
\alias{stationary_particles_map}
\title{Map stationary particles}
\usage{
stationary_particles_map(input, subset = 100)
}
\arguments{
\item{input}{input particle tracks (e.g. coralseed:::Mermaid_PointSource_Bay_01)}

\item{subset}{n particles to subset}
}
\description{
Function to identify and map stationary particles in particle tracks
}
