% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralseed-package.R
\docType{package}
\name{coralseed-package}
\alias{coralseed}
\alias{coralseed-package}
\title{coralseed: What the Package Does (One Line, Title Case)}
\description{
a spatially explicit individual-based probabilistic model for predicting coral settlement following larval dispersal on coral reefs
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
