% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralseed-package.R
\docType{package}
\name{coralseed-package}
\alias{coralseed}
\alias{coralseed-package}
\title{coralseed: coralseed: a spatially explicit reef-scale model of larval reseeding}
\description{
a spatially explicit individual-based probabilistic model for predicting coral settlement following larval dispersal on coral reefs
}
\author{
\strong{Maintainer}: George Roff \email{george.roff@csiro.ai} (\href{https://orcid.org/0000-0002-2823-0575}{ORCID})

}
\keyword{internal}
