% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralseed-package.R
\docType{package}
\name{coralseed-package}
\alias{coralseed}
\alias{coralseed-package}
\title{coralseed: Spatially explicit probabilistic modelling of coral settlement}
\description{
coralseed: a spatially explicit reef-scale model of coral settlement following larval reseeding
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marine-ecologist.github.io/coralseed/}
}

}
\author{
\strong{Maintainer}: George Roff \email{george.roff@csiro.au}

}
\keyword{internal}
