% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R, R/extract_parallel.R,
%   R/map_coralseed.R
\name{extract_parallel}
\alias{extract_parallel}
\alias{seascapecolors}
\alias{map_coralseed}
\title{Color palettes}
\usage{
seascapecolors()

extract_parallel(input, type)

map_coralseed(
  seed_particles_input = particles,
  settle_particles_input = settlers,
  seascape_probability = seascape,
  restoration.plot = c(100, 100),
  show.tracks = TRUE,
  show.footprint = FALSE,
  subsample = NULL,
  scalebar = 200,
  webGL = FALSE
)
}
\arguments{
\item{input}{list of \%dopar\% outputs}

\item{type}{either points or paths}

\item{seed_particles_input}{input from seed_particles_input}

\item{settle_particles_input}{input from settle_particles_input}

\item{seascape_probability}{input from seascape_probability}

\item{restoration.plot}{dimensions of the restoration plot in metres}

\item{show.tracks}{option to show particle tracks (TRUE, will be large files and slower renders) or without tracks (FALSE, smaller files and faster renders)}

\item{scalebar}{set scale to X metres}
}
\description{
Functions to colour maps

Function to extract points and paths from multiple \%dopar\% outputs

Function to quickly map settlers
}
