% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parallel.R, R/map_coralseed.R
\name{extract_parallel}
\alias{extract_parallel}
\alias{map_coralseed}
\title{Settlement statistics}
\usage{
extract_parallel(input, type, ...)

map_coralseed(
  seed_particles = particles,
  settle_particles = settlers,
  seascape_probability = seascape,
  restoration.plot = c(100, 100),
  show.tracks = TRUE
)
}
\arguments{
\item{input}{list of \%dopar\% outputs}

\item{type}{either points or paths}

\item{...}{pass arguments}

\item{seed_particles}{input from seed_particles}

\item{settle_particles}{input from settle_particles}

\item{seascape_probability}{input from seascape_probability}

\item{restoration.plot}{dimensions of the restoration plot in metres}

\item{show.tracks}{option to show particle tracks (TRUE, will be large files and slower renders) or without tracks (FALSE, smaller files and faster renders)}
}
\description{
Function to extract points and paths from multiple \%dopar\% outputs

Function to quickly map settlers
}
