% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_silent.R
\name{ggplot_silent}
\alias{ggplot_silent}
\title{Silence ggplot warnings}
\usage{
ggplot_silent(plot_obj, ...)
}
\arguments{
\item{plot_obj}{input}

\item{...}{Arguments are passed to \code{print.ggplot()}}
}
\description{
Replacement for \code{print.ggplot()} that mutes warnings.
}
