% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coralseed_future.R
\name{map_coralseed_future}
\alias{map_coralseed_future}
\title{Map coralseed settlers subfunction}
\usage{
map_coralseed_future(
  points = output_points,
  paths = paths,
  seascape_probability = seascape,
  restoration.plot = c(100, 100),
  cellsize = 10,
  concavity = 1.2,
  length_threshold = 10
)
}
\arguments{
\item{points}{input from seed_particles}

\item{paths}{input from settle_particles}

\item{seascape_probability}{input from seascape_probability}

\item{restoration.plot}{dimensions of the restoration plot in metres}

\item{cellsize}{grid cell width/height}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{length_threshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes.}
}
\description{
Function to quickly map settlers for multiple
}
