% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_draws.R
\name{plot_draws}
\alias{plot_draws}
\title{plot draws}
\usage{
plot_draws(input, slice = 100)
}
\arguments{
\item{slice}{n number of simulations for spaghetti plot}

\item{n}{number of simulations}
}
\description{
Function to plot spaghetti draws from time-to-competency brms model
}
