% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_settlers.R
\name{plot_settlers}
\alias{plot_settlers}
\title{Plot output from settle_particles}
\usage{
plot_settlers(input = NULL, pch = 16, size = 0.5, ...)
}
\arguments{
\item{input}{input}

\item{pch}{pch}

\item{size}{size}

\item{...}{passes functions}
}
\description{
Function to plot output of seed_particles in base R for quick cheks
}
\examples{
# example code
# particles <- seed_particles("mermaid", seascape=seascape, limit_time=720, competency.function = "exponential", simulate.mortality = "typeI", simulate.mortality.n = 0.1, probability="additive",return.plot=FALSE, silent=TRUE)
# plot_particles(particles)
}
