% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_competency.R
\name{predict_competency}
\alias{predict_competency}
\title{Predict_competency}
\usage{
predict_competency(
  n_id,
  n_sims = 1000,
  competency.function = "exponential",
  b_Intercept = b_Intercept,
  sort = TRUE,
  seed.value = NULL,
  return.plot = TRUE,
  ...
)
}
\arguments{
\item{n_id}{input n_id (number of particles)}

\item{n_sims}{number of sims for randomised datasets (see plot for traces)}

\item{competency.function}{distribution, one of "weibull", "exponential", "lognormal"}

\item{sort}{sort output by ID, otherwise if FALSE randomly distribute}

\item{seed.value}{set seed value, defaults to NULL}

\item{return.plot}{return output}

\item{...}{passes functions}
}
\description{
Function to generate probability distribution from Allen Coral Atlas input files
predict_competency(n_id=100, nsims=1000, competency.function="exponential")
}
