% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_competency.R
\name{predict_competency}
\alias{predict_competency}
\title{Predict_competency}
\usage{
predict_competency(
  n_id,
  competency.function = "exponential",
  sort = FALSE,
  set.seed = NULL,
  return.plot = TRUE,
  ...
)
}
\arguments{
\item{n_id}{input n_id}

\item{competency.function}{distribution, one of "weibull", "exp", "log"}

\item{sort}{sort output by ID, otherwise if FALSE randomly distribute}

\item{set.seed}{set seed value, defaults to NULL}

\item{return.plot}{return output}

\item{...}{passes functions}
}
\description{
Function to generate probability distribution from Allen Coral Atlas input files
}
