% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape_probability.R
\name{seascape_probability}
\alias{seascape_probability}
\title{Seascape Probability}
\usage{
seascape_probability(
  reefoutline = NULL,
  habitat = NULL,
  probability = NULL,
  ...
)
}
\arguments{
\item{reefoutline}{geojson or shp files for Coral Atlas benthic habitat maps for reef outline}

\item{habitat}{habitat or shp files for Coral Atlas geomorphic maps}

\item{probability}{habitat specific probabilities (mean and SE) defined as in example below. Defaults to NULL parameterisation embedded in function (see example below)}

\item{...}{passes functions}
}
\description{
Function to generate probability distribution from Allen Coral Atlas input files
}
