% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_particles.R
\name{seed_particles}
\alias{seed_particles}
\title{Seed particles}
\usage{
seed_particles(
  input = NULL,
  example = NULL,
  seascape = NULL,
  subsample = NULL,
  simulate.mortality = "none",
  simulate.mortality.n = 0.1,
  competency.function = "exp",
  limit.time = NA,
  set.centre = TRUE,
  set.seed = NULL,
  silent = FALSE,
  return.plot = FALSE,
  ...
)
}
\arguments{
\item{input}{input}

\item{example}{example}

\item{seascape}{shp file inputs from seascape_probability()}

\item{subsample}{subsample to n samples}

\item{simulate.mortality}{set mortality type via simulate_mortality() one of "typeI","typeII", "typeIII" (defaults to "none")}

\item{simulate.mortality.n}{set proportion of corals to kill over a 24hr period, where 0 is none, 1 is 100 (defaults to 0.1 or 10\%)}

\item{competency.function}{set distribution to define competency from time-to-settlement model, one of "weibull", "exp", "log" (defaults to "exp")}

\item{limit.time}{limit the time series, for example 720 will limit the settlement results between 0-12hrs (defaults to "NA")}

\item{set.centre}{reset CONNIE input to have a central t0 point (defaults to TRUE)}

\item{set.seed}{set seed for consistent results (defaults to NULL)}

\item{silent}{silence printing results while running (defaults to FALSE)}

\item{return.plot}{return outputs for seed_particles (defaults to "FALSE")}

\item{...}{passes functions}
}
\description{
Function to seed particles
}
\details{
seed_particles() uses predict_competency(), simulate_mortality()
}
