% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restoration_plot.R
\name{set_restoration_plot}
\alias{set_restoration_plot}
\title{Set restoration plot}
\usage{
set_restoration_plot(
  input = NULL,
  width = NULL,
  length = NULL,
  crs = 4326,
  center = "coralseed"
)
}
\arguments{
\item{input}{input \code{sf} object used to derive the center}

\item{width}{width in metres}

\item{length}{length in metres}

\item{crs}{CRS for output object (default 4326)}

\item{center}{either "coralseed" for t0-based centroid, or "sf" for direct centroid}
}
\description{
Function to create a rectangular polygon around the release point (centroid or t0 of particle release) for plotting in tmap
}
