% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settle_particles.R
\name{settle_particles}
\alias{settle_particles}
\title{Settle Particles}
\usage{
settle_particles(
  input,
  probability = "additive",
  silent = TRUE,
  sample.n = NULL,
  seed.value = NULL,
  ...
)
}
\arguments{
\item{input}{input (defaults to particle_points_expanded from seed_particles() function)}

\item{probability}{one of "additive", "rapid", "lagged")}

\item{silent}{silence outputs}

\item{sample.n}{subsample to n particles}

\item{...}{pass arguments}

\item{set.seed}{seed}
}
\description{
Function to determine probability of settlement if particles pass suitable substrate
options: additive (if p=1, settle somewhere in that habitat along the particle trajectory regardless of time)
lagged (if p=1, settle somewhere in the habitat in the first 10 minutes)
rapid (if p=1, settle immediately habitat on trajectory )
}
