% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{Simulate Mortality}
\usage{
simulate_mortality(
  input = NULL,
  simulate.mortality = NULL,
  simulate.mortality.n = 0.1,
  return.plot = TRUE,
  silent = FALSE,
  seed.value = NULL,
  ...
)
}
\arguments{
\item{input}{input (e.g. from \code{seed_particles()}, usually a time-expanded \code{sf} object)}

\item{simulate.mortality}{select mortality type ("typeI", "typeII", or "typeIII")}

\item{simulate.mortality.n}{proportion of population to simulate mortality on (0–1)}

\item{return.plot}{generate plot output (default: TRUE)}

\item{silent}{suppress printed messages (default: FALSE)}

\item{seed.value}{seed for reproducibility (default: NULL)}

\item{...}{pass additional args}
}
\description{
Function to simulate mortality of a population. \code{simulate.mortality.n} is the proportion of the population to simulate mortality over a 24hr period.
}
\details{
Mortality types I, II, III are Weibull-distributed. Set \code{return.plot=TRUE} to visualise simulation vs empirical.

Example:
tmp <- simulate_mortality(particle_points_expanded, simulate.mortality = "typeIII", simulate.mortality.n = 0.2, return.plot = TRUE)
}
